Insert and Dump are working according to spec; read notes pls
Major changes: all functions are now completely recursive; no node knows
about its coordinates, width, or depth. The only data structure with
knowledge of coordinates are the individual points within the
LinkedList. FlyLeaf was made as a static final member of the Tree, is
accessed in the other nodes with a Class.getter. Insert and dump are
working. Removed the toString methods to make it easier to work with
recursion in both cases. 

Minor Changes: rearranged the traversal and output order to match the
output file syntax. Added parenthesis to the toString of the Points to
match spec. Removed the unused interface methods to help out with our
test coverage until we need em. 

To-Do: The Milestone was pushed to Saturday, but if we can get our Test
Coverage to like 90-95% by tomorrow night, that would be great. Need to
appease the Web-CAT gods and do commenting and stuff.