at 75% problem coverage, 78% testing
added the duplicates method correctly. adjusted a few flaws in the
insert and adjustTree methods. for some reason i wrote it so the leaf
would only split at size 4, but not greater. and the tree splitting
wasnt done quite right. duplicates works by finding the leaves, looping
through the linkedlist for duplicates. Search was added too, but it was
just changing output formates in the CommandParser to pass the webcat
tests. nothing added to tree because we search by name, which is a
skiplist function. need to write JUnit tests and comment more.
RegionSearch and remove are the only two methods left.