Remove working to fit Web-CAT tests
a lot of cleaning up is in order, but I got remove to work. It's
probably really inefficient, but it works. Basically removes the right
value, recurses back to the top of the tree, then goes back down and
adjusts. will try to work it so it fixes itself on the way back up, but
for now this will do. Need comments and tests, and a lot of cleaning up
of redundant and unnecessary code. (p.s. this is my way of
procrastinating on studying for tests. Coding > combo)